pyproject.toml
[tool.poetry]
name = "fastzdp-login"
version = "0.1.3"
description = "专为Fast API打造的用户登录注册相关功能的微代码框架"
authors = ["zhangdapeng"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlmodel = {extras = ["mysql"], version = "^0.0.19"}
pymysql = "^1.1.1"
cryptography = "^42.0.8"
passlib = "^1.7.4"
pyjwt = "^2.8.0"


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
